Consolidate preflight logic
